+++ crates/verdin/src/lib.rs
@@ -6,4 +6,5 @@ pub mod cli;
 pub mod config;
 mod new;
 pub mod server;
+mod typescript;
 mod uploads;